fix: add float number for weight and price

+++ store/server/models/mapping.js
@@ -21,11 +21,11 @@ const BasketProduct = sequelize.define('basket_product', {
 const Product = sequelize.define('product', {
     id: {type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     name: {type: DataTypes.STRING, unique: true, allowNull: false},
-    price: {type: DataTypes.INTEGER, allowNull: false},
+    price: {type: DataTypes.FLOAT, allowNull: false},
     rating: {type: DataTypes.INTEGER, defaultValue: 0},
     image: {type: DataTypes.STRING, allowNull: false},
     article: {type: DataTypes.INTEGER, unique: true, allowNull: false},
-    weight: {type: DataTypes.INTEGER, allowNull: false},
+    weight: {type: DataTypes.FLOAT, allowNull: false},
 })
 
 const Category = sequelize.define('category', {
